% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_anes}
\alias{data_anes}
\title{American National Election Studies 2022 pilot study data}
\format{
A data frame with 1,307 rows and 146 columns:
\describe{
  \item{party}{party reported as \code{pid3} in the raw data}
  \item{follow}{response to "Would you say you follow what’s going on in government and public affairs most of the time, some of the time, only now and then, or hardly at all?"}
  \item{reg}{response to "Are you registered to vote, or not?"}
  ...
}
}
\source{
\url{https://electionstudies.org/data-center/}
}
\usage{
data_anes
}
\description{
A subset of data from the American National Election Studies 2022 Pilot Study
}
\keyword{datasets}
