% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_gen.R
\name{data_gen_loc}
\alias{data_gen_loc}
\title{Generates locally dependent Bernoulli data}
\usage{
data_gen_loc(N, J, K, rho, block, alphas_pi, pars = c(0.2, 0.2))
}
\arguments{
\item{N}{integer. The number of subjects}

\item{J}{integer. The number of items}

\item{K}{integer. The number of extreme latent profiles}

\item{rho}{numeric. The auto-regressive parameter for local dependence. Needs to be in \code{[0,1]}}

\item{block}{integer. The local dependence block size}

\item{alphas_pi}{vector of numeric values. Specifies the Dirichlet parameters for membership score generation. Need to be length of \code{K}}

\item{pars}{vector of non-negative numeric values. Specifies the Beta distribution shape parameters for item parameter generation. Need to be length of \code{2}}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{Pi} --- Numeric matrix. Membership score matrix
\item \code{Theta} --- Numeric matrix. Flattened item parameter matrix
\item \code{R} --- Binary matrix of size \code{N}x\code{J}. Generated locally dependent data matrix 
}
}
\description{
This function generates locally dependent Bernoulli distributed GoM data.
}
