% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_gen.R
\name{data_gen_pois}
\alias{data_gen_pois}
\title{Generates Poisson data}
\usage{
data_gen_pois(N, J, K, alphas_pi, pars = c(1, 2))
}
\arguments{
\item{N}{integer. The number of subjects}

\item{J}{integer. The number of items}

\item{K}{integer. The number of extreme latent profiles}

\item{alphas_pi}{vector of numeric values. Specifies the Dirichlet parameters for membership score generation. Need to be length of \code{K}}

\item{pars}{vector of positive numeric values. Specifies the Gamma distribution shape and rate parameters for item parameter generation. Need to be length of \code{2}}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{Pi} --- Numeric matrix. Membership score matrix
\item \code{Theta} --- Numeric matrix. Flattened item parameter matrix
\item \code{R} --- Binary matrix of size \code{N}x\code{J}. Generated Poisson data matrix 
}
}
\description{
This function generates Poisson distributed GoM data.
}
