% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_gen.R
\name{data_gen_pol}
\alias{data_gen_pol}
\title{Generates polytomous data}
\usage{
data_gen_pol(N, J, K, C, alphas_pi, alphas_theta)
}
\arguments{
\item{N}{integer. The number of subjects}

\item{J}{integer. The number of items}

\item{K}{integer. The number of extreme latent profiles}

\item{C}{integer. The number of possible outcomes for each item}

\item{alphas_pi}{vector of numeric values. Specifies the Dirichlet parameters for membership score generation. Need to be length of \code{K}
#' @param alphas_theta vector of numeric values. Specifies the Dirichlet parameters for item parameter generation. Need to be length of \code{C}}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{Pi} --- Numeric matrix. Membership score matrix
\item \code{Theta} --- Numeric matrix. Flattened item parameter matrix
\item \code{R_flattened} --- Binary matrix of size \code{N}x(\code{CJ}). Generated flattened data matrix 
\item \code{R_pol} --- Integer matrix of size \code{N}x\code{J}. Generated polytomous data matrix 
}
}
\description{
This function generates polytomous GoM data
}
