% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_senate}
\alias{data_senate}
\title{112th Senate Roll Call Data}
\format{
A data frame with 94 rows and 487 columns. Missing responses are imputed with average response rate.
\describe{
  \item{party}{party information: Dem. or Repub.}
  \item{roll1-486}{binary response to 486 roll calls}
  ...
}
}
\source{
\url{https://legacy.voteview.com/senate112.htm}
}
\usage{
data_senate
}
\description{
A subset of binary data from the 112th Senate roll call data
}
\keyword{datasets}
