% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DhLCM.R
\name{DhLCM}
\alias{DhLCM}
\title{DhLCM clustering and estimation}
\usage{
DhLCM(
  R,
  K,
  spectral = "heteroPCA",
  norm = "L2",
  dist = "Bern",
  T0 = 20,
  nstart = 10,
  S0 = NULL,
  clustering_only = F
)
}
\arguments{
\item{R}{Numeric matrix. Data matrix.}

\item{K}{Positive integer. The number of top eigenvectors/left singular vectors to be extracted.}

\item{spectral}{Numeric matrix or character. One of data matrix, \code{"heteroPCA"} and \code{"SVD"}. 
If is a matrix, it is treated as U. Otherwise needs to be a string that 
specifies the method to be used to obtain the top \code{K} 
eigenvectors/left singular vectors. 
\code{"heteroPCA"} implements the heteroPCA method. 
\code{"SVD"} performs ordinary singular vector decomposition.}

\item{norm}{Character or \code{NULL}. One of \code{"L2"}, \code{"L1"}, \code{"SCORE"}, and \code{NULL}. 
Specifies the method to be used for normalization on the eigenvectors/left singular vectors. 
\code{"L2"} performs L2 normalization. 
\code{"L1"} performs L1 normalization.
\code{"SCORE"} performs SCORE normalization.
\code{"NA"} does not perform normalization.}

\item{dist}{Character. One of \code{"Bern"}, \code{"Binom"}, and \code{"Pois"}. Specifies the data distribution.
\code{"Bern"} assumes the Bernoulli distribution.
\code{"Binom"} assumes the Binomial distribution.
\code{"Pois"} assumes the Poisson distribution.}

\item{T0}{Positive integer. The number of iterations for heteroPCA. Only used when spectral is \code{'heteroPCA'}}

\item{nstart}{Positive integer. The number of initial starts in the kmeans function.}

\item{S0}{Vector or \code{NULL}. If is not \code{NULL}, used to permute the labels.}

\item{clustering_only}{Boolean. When \code{true}, only clustering is conducted.}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{U} --- Numeric matrix. Estimation of the left singular matrix.
\item \code{T_hat} --- Numeric matrix. Estimation of the \eqn{\Theta}{Theta} matrix.
\item \code{sigma2_hat} --- Numeric vector (>=0). Asymptotic variance for each element of \code{T_hat}.
\item \code{S_hat} --- Numeric vector. Clustered membership for each subject.
\item \code{Z_hat} --- Numeric matrix. Clustered membership for each subject in binary matrix form.
}
}
\description{
This function performs k-means clustering on the top \code{K} eigenvectors/left singular vectors, and estimates the DhLCM model parameters
}
\references{
Lyu, Zhongyuan, Ling Chen, and Yuqi Gu. "Degree-heterogeneous Latent Class Analysis for High-dimensional Discrete Data." arXiv preprint arXiv:2402.18745 (2024).
}
