% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_deletion.R
\name{diag_deletion}
\alias{diag_deletion}
\title{Diagonal deletion}
\usage{
diag_deletion(X)
}
\arguments{
\item{X}{Numeric matrix}
}
\value{
A matrix of with diagonals set to 0
}
\description{
This function takes in a matrix, and returns the diagonal-deleted matrix
}
