% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGoM.R
\name{gGoM}
\alias{gGoM}
\title{Generalized-GoM estimation}
\usage{
gGoM(
  R,
  K,
  pol = T,
  dist = NULL,
  large = T,
  prune = T,
  r = 10,
  q = 0.4,
  e = 0.2
)
}
\arguments{
\item{R}{data matrix.}

\item{K}{integer. The number of extreme latent profiles. \code{K} should be at least 2.}

\item{pol}{logical; if true, assume GoM model with polytomous response, and flattening is applied. Item parameter estimation \code{T_hat} is also flattened.}

\item{dist}{character; One of \code{"Bern"}, \code{"Binom"}, and \code{"Pois"}. Specifies the data distribution.
\code{"Bern"} assumes the Bernoulli distribution.
\code{"Binom"} assumes the Binomial distribution.
\code{"Pois"} assumes the Poisson distribution.}

\item{large}{logical; if true, \code{K} needs to be at least 3 and use the large-scale SVD function \code{RSpectra::svds}.}

\item{prune}{logical; if true, the pruning step is performed.}

\item{r}{the number of neighbors to consider in pruning. Used only when \code{prune} is \code{TRUE}. Default value is 10.}

\item{q}{the cutoff for the upper quantile of row norms. Used only when \code{prune} is \code{TRUE}. Higher \code{q} leads to more points being pruned. Default value is 0.4.}

\item{e}{the cutoff for the upper quantile of average distance. Used only when \code{prune} is \code{TRUE}. Higher \code{e} leads to more points being pruned. Default value is 0.2.}

\item{lower}{the minimum value for item parameters. Default value is 0.}

\item{upper}{the minimum value for item parameters. Default value is 1.}
}
\value{
The function returns a list with the following components:
\itemize{
\item{\code{P_hat} the estimated membership scores.}
\item{\code{T_hat} the estimated item response parameters.}
\item{\code{R_hat} the estimated response expectation.}
\item{\code{S_hat} the estimated indices of pure subjects.}
\item{\code{t} computation time.}
}
}
\description{
Estimation algorithm for generalized-GoM models
}
\references{
Chen, Ling, and Yuqi Gu. "A spectral method for identifiable grade of membership analysis with binary responses." Psychometrika (2024): 1-32.

Chen, Ling, Chengzhu Huang, and Yuqi Gu. "Generalized Grade-of-Membership Estimation for High-dimensional Locally Dependent Data." arXiv preprint arXiv:2412.19796 (2024).
}
