% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroPCA.R
\name{heteroPCA}
\alias{heteroPCA}
\title{HeteroPCA}
\usage{
heteroPCA(R, K, T0)
}
\arguments{
\item{R}{Numeric matrix. The matrix to perform HeteroPCA.}

\item{K}{Positive integer. The number of top eigenvectors to be extracted.}

\item{T0}{Positive integer. The number of iterations.}
}
\value{
Numeric matrix \code{U_hat}
}
\description{
This function implements the HeteroPCA algorithm
}
\references{
Zhang, Anru R., T. Tony Cai, and Yihong Wu. "Heteroskedastic PCA: Algorithm, optimality, and applications." The Annals of Statistics 50.1 (2022): 53-80.
}
