% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.R
\name{perm}
\alias{perm}
\title{perm}
\usage{
perm(x, p)
}
\arguments{
\item{x}{Numeric vector. Vector of labels with integer values 1, ..., K}

\item{p}{Numeric vector. An integer permutation vector.}
}
\value{
Permuted vector \code{x_perm}
}
\description{
This function performs permutation
}
