% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{pruning}
\alias{pruning}
\title{pruning}
\usage{
pruning(mat, r = 10, q = 0.4, e = 0.2)
}
\arguments{
\item{mat}{a numeric matrix to be pruned.}

\item{r}{the number of neighbors to consider. Default value is 10.}

\item{q}{the cutoff for the upper quantile of row norms. Higher `q` leads to more points being pruned. Default value is 0.4.}

\item{e}{the cutoff for the upper quantile of average distance. Higher `e` leads to more points being pruned. Default value is 0.2.}
}
\value{
\describe{
  \item{\code{indices} the index vector of the rows to be pruned from the left singular matrix.}
}
}
\description{
Locate noisy points in the data simplex.
}
\references{
Mao, X., Sarkar, P., & Chakrabarti, D. (2021). Estimating mixed memberships with sharp eigenvector deviations. Journal of the American Statistical Association, 116(536), 1928-1940.
}
