% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{rescale_T}
\alias{rescale_T}
\title{rescale_T}
\usage{
rescale_T(T_mat, Cs)
}
\arguments{
\item{T_mat}{Numeric matrix. Item parameter matrix.}

\item{Cs}{Integer vector. The number of categories for each item}
}
\value{
\describe{
  \item{\code{T_mat}    flattened item parameter matrix estimation.}
}
}
\description{
Re-scale the item parameter estimation \code{T_hat} for polytomous GoM
}
