% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa.R
\name{spa}
\alias{spa}
\title{spa}
\usage{
spa(mat)
}
\arguments{
\item{mat}{the (pruned) left singular matrix to conduct SPA on.}
}
\value{
\describe{
  \item{\code{S_hat} a vector of the pure subject indices.}
}
}
\description{
A sequential projection algorithm (SPA) to find the pure subjects
}
\references{
Gillis, N. and Vavasis, S. A. (2013). Fast and robust recursive algorithms for separable nonnegative matrix factorization. IEEE Transactions on Pattern Analysis and Machine Intelligence, 36(4):698–714.
}
