% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{flatten}
\usage{
flatten(R)
}
\arguments{
\item{R}{integer matrix. The polytomous response data matrix.}
}
\value{
\describe{
  \item{\code{R_flattened}    flattened binary matrix.}
}
}
\description{
Flatten the polytomous matrix to a fat binary matrix
}
